Add support for inline mode
