+++ addon/components/bootstrap-datepicker-inline.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import DatepickerSupport from 'ember-cli-bootstrap-datepicker/components/datepicker-support';
+
+export default Ember.Component.extend(DatepickerSupport, {
+  tagName: 'div'
+});
