+++ app/components/bootstrap-datepicker-inline.js
@@ -1,0 +1,4 @@
+import Ember from 'ember';
+import BootstrapDatepickerInlineComponent from 'ember-cli-bootstrap-datepicker/components/bootstrap-datepicker-inline';
+
+export default BootstrapDatepickerInlineComponent;
