+++ tests/dummy/app/styles/app.css
@@ -9,7 +9,7 @@ body {
   padding-bottom: 20px;
 }
 
-.panel-footer pre {
+.panel-heading pre {
   padding: 0;
   margin-top: 0;
   margin-bottom: 0;
